refactor: extract `tests/wallets/helpers.py`
